+++ modules/SofaConstraint/GenericConstraintSolver.cpp
@@ -20,22 +20,11 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 
-#include <sofa/helper/logging/Messaging.h>
-
 #include <SofaConstraint/GenericConstraintSolver.h>
 #include <sofa/core/visual/VisualParams.h>
-
-#include <sofa/simulation/MechanicalVisitor.h>
-#include <sofa/simulation/SolveVisitor.h>
 #include <sofa/simulation/VectorOperations.h>
-
-#include <sofa/simulation/Simulation.h>
 #include <sofa/helper/AdvancedTimer.h>
-#include <sofa/helper/system/thread/CTime.h>
-#include <math.h>
-
 #include <sofa/core/ObjectFactory.h>
-
 #include "ConstraintStoreLambdaVisitor.h"
 
 namespace sofa
@@ -212,8 +201,6 @@ bool GenericConstraintSolver::buildSystem(const core::ConstraintParams *cParams,
     // mechanical action executed from root node to propagate the constraints
     simulation::MechanicalResetConstraintVisitor(cParams).execute(context);
     // calling buildConstraintMatrix
-	//simulation::MechanicalAccumulateConstraint(&cparams, cParams->j(), numConstraints).execute(context);
-
 	simulation::MechanicalBuildConstraintMatrix(cParams, cParams->j(), numConstraints).execute(context);
     simulation::MechanicalAccumulateMatrixDeriv(cParams, cParams->j(), reverseAccumulateOrder.getValue()).execute(context);
 
@@ -279,7 +266,7 @@ bool GenericConstraintSolver::buildSystem(const core::ConstraintParams *cParams,
             }
 
             if (!foundCC)
-                serr << "WARNING: no constraintCorrection found for constraint" << c_id << sendl;
+                msg_error() << "WARNING: no constraintCorrection found for constraint" << c_id ;
 
             double** w =  current_cp->getW();
             for(unsigned int m = c_id; m < c_id + l; m++)
@@ -310,7 +297,6 @@ bool GenericConstraintSolver::buildSystem(const core::ConstraintParams *cParams,
         sofa::helper::AdvancedTimer::stepEnd  ("Get Compliance");
         msg_info() << " computeCompliance_done "  ;
     }
-
 
     if ( displayTime.getValue() )
     {
@@ -506,7 +492,7 @@ bool GenericConstraintSolver::applyCorrection(const core::ConstraintParams *cPar
 
 	if (displayTime.getValue())
     {
-        sout << " TotalTime " << ((double) timerTotal.getTime() - timeTotal) * timeScale << " ms" << sendl;
+        msg_info() << " TotalTime " << ((double) timerTotal.getTime() - timeTotal) * timeScale << " ms";
     }
 
     return true;
@@ -545,7 +531,7 @@ void GenericConstraintSolver::lockConstraintProblem(sofa::core::objectmodel::Bas
         }
     }
     // All constraint problems are locked
-    serr << "All constraint problems are locked, request from " << (from ? from->getName() : "NULL") << " ignored" << sendl;
+    msg_error() << "All constraint problems are locked, request from " << (from ? from->getName() : "NULL") << " ignored";
 }
 
 void GenericConstraintProblem::clear(int nbC)
@@ -1051,7 +1037,7 @@ void GenericConstraintProblem::unbuiltGaussSeidel(double timeout, GenericConstra
                 break;
             }
         }
-        else if(error < tol/* && i>0*/) // do not stop at the first iteration (that is used for initial guess computation)
+        else if(error < tol)
         {
             convergence = true;
             break;
@@ -1101,6 +1087,56 @@ void GenericConstraintProblem::unbuiltGaussSeidel(double timeout, GenericConstra
     }
 }
 
+sofa::core::MultiVecDerivId GenericConstraintSolver::getLambda()  const
+{
+    return m_lambdaId;
+}
+
+sofa::core::MultiVecDerivId GenericConstraintSolver::getDx() const
+{
+    return m_dxId;
+}
+
+MechanicalGetConstraintResolutionVisitor::MechanicalGetConstraintResolutionVisitor(const core::ConstraintParams* params, std::vector<core::behavior::ConstraintResolution*>& res)
+: simulation::BaseMechanicalVisitor(params)
+, cparams(params)
+, _res(res)
+, _offset(0)
+{
+#ifdef SOFA_DUMP_VISITOR_INFO
+  setReadWriteVectors();
+#endif
+}
+
+MechanicalGetConstraintResolutionVisitor::Result MechanicalGetConstraintResolutionVisitor::fwdConstraintSet(simulation::Node* node, core::behavior::BaseConstraintSet* cSet)
+{
+  if (core::behavior::BaseConstraint *c=cSet->toBaseConstraint())
+  {
+    ctime_t t0 = begin(node, c);
+    c->getConstraintResolution(cparams, _res, _offset);
+    end(node, c, t0);
+  }
+  return RESULT_CONTINUE;
+}
+
+/// Return a class name for this visitor
+/// Only used for debugging / profiling purposes
+const char* MechanicalGetConstraintResolutionVisitor::getClassName() const
+{
+    return "MechanicalGetConstraintResolutionVisitor";
+}
+
+bool MechanicalGetConstraintResolutionVisitor::isThreadSafe() const
+{
+    return false;
+}
+
+bool MechanicalGetConstraintResolutionVisitor::stopAtMechanicalMapping(simulation::Node* node, core::BaseMapping* map)
+    {
+        SOFA_UNUSED(node);
+        SOFA_UNUSED(map);
+        return false;
+    }
 
 
 int GenericConstraintSolverClass = core::RegisterObject("A Generic Constraint Solver using the Linear Complementarity Problem formulation to solve Constraint based components")
